+++ templates/messages/index.html
@@ -7,8 +7,6 @@
 
 <div class="container">
    {% if messages %}
-
-   
     <div class="row">
         
         <div class="col-md-6">
@@ -18,11 +16,17 @@
                 <div class="panel-body">
                     <ul class="chat">
                         {% for message in messages%}
+                            {% if message[6] != session['id'] %}
                         
-                        <li class="left clearfix"><span class="chat-img pull-left">
-                            <img src="http://placehold.it/50/55C1E7/fff&text=U" alt="User Avatar" class="img-circle" />
-                        </span>
-                            <div class="chat-body clearfix">
+                                <li class="left clearfix"><span class="chat-img pull-left">
+                                    <img src="http://placehold.it/50/55C1E7/fff&text=U" alt="User Avatar" class="img-circle" />
+                                </span>
+                            {% else %}
+                                <li class="right clearfix"><span class="chat-img pull-right">
+                                    <img src="http://placehold.it/50/FA6F57/fff&text=ME" alt="User Avatar" class="img-circle" />
+                                </span>
+                            {% endif %}   
+                                <div class="chat-body clearfix">
                                 <div class="header">
                                     <strong class="primary-font"> {{message[1]}} Topic:  
                                     {{message[3]}}
